requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astropy==6.1.4
astropy-iers-data==0.2024.9.30.0.32.59
astroquery==0.4.7
asttokens==3.0.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.2.0
boto3==1.34.7
botocore==1.34.7
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.3.1
cors==1.0.1
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.11
daphne==4.0.0
debugpy==1.8.11
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
executing==2.1.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
freetype-py==2.3.0
fsspec==2025.2.0
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
immutabledict==4.2.1
incremental==22.10.0
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.4.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lxml==4.9.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.2
openai==1.62.0
opencv-python==4.10.0.84
openpyxl==3.1.5
ortools==9.11.4210
oscrypto==1.3.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
Pillow==10.1.0
platformdirs==4.2.0
preshed==3.0.9
prettytable==3.12.0
prompt_toolkit==3.0.48
protobuf==5.26.1
psutil==6.1.1
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.25.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyerfa==2.0.1.4
Pygments==2.19.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyOpenSSL==23.3.0
pyparsing==3.2.1
pypdf==3.17.3
PyPDF2==3.0.1
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2024.2
pyvo==1.5.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==7.4.2
regex==2024.11.6
reportlab==4.0.8
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rlPyCairo==0.3.0
s3transfer==0.10.0
scipy==1.15.0
seaborn==0.13.2
service-identity==23.1.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.5.1
stack-data==0.6.3
svglib==1.5.1
sympy==1.13.1
thinc==8.3.4
tinycss2==1.2.1
tldextract==5.1.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.3
virtualenv==20.25.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.6.0
wrapt==1.17.2
xhtml2pdf==0.2.13
XlsxWriter==3.2.2
xlwt==1.3.0
zope.event==5.0
zope.interface==7.0.3
